NIFI-7592: Rename MiNiFi references to Headless NiFi, added profile to build headless assembly
