+++ nifi-assembly/pom.xml
@@ -943,6 +943,28 @@ language governing permissions and limitations under the License. -->
                 </dependency>
             </dependencies>
     	</profile>
+        <profile>
+            <id>headless</id>
+            <!-- This profile handles the inclusion of nifi-server artifacts, such as the Web/UI NAR(s). -->
+            <activation>
+                <activeByDefault>true</activeByDefault>
+            </activation>
+            <dependencies>
+                <dependency>
+                    <groupId>org.apache.nifi</groupId>
+                    <artifactId>nifi-server-nar</artifactId>
+                    <version>1.12.0-SNAPSHOT</version>
+                    <type>nar</type>
+                    <scope>provided</scope>
+                </dependency>
+                <dependency>
+                    <groupId>org.apache.nifi</groupId>
+                    <artifactId>nifi-headless-server-nar</artifactId>
+                    <version>1.12.0-SNAPSHOT</version>
+                    <type>nar</type>
+                </dependency>
+            </dependencies>
+        </profile>
 	<profile>
             <id>include-accumulo</id>
             <!-- This profile handles the inclusion of nifi-accumulo artifacts. -->
