+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/pom.xml
@@ -48,7 +48,7 @@
         <module>nifi-mock-authorizer</module>
         <module>nifi-shell-authorizer</module>
         <module>nifi-stateless</module>
-        <module>minifi-server</module>
+        <module>nifi-headless-server</module>
     </modules>
     <dependencies>
         <dependency>
