+++ nifi-nar-bundles/nifi-framework-bundle/nifi-headless-server-nar/src/main/resources/META-INF/NOTICE
@@ -1,4 +1,4 @@
-minifi-server-nar
+nifi-headless-server-nar
 Copyright 2014-2020 The Apache Software Foundation
 
 This product includes software developed at
